+++ Js2IL.Tests/Hosting/JavaScript/Hosting_TypedExports_Async.js
@@ -1,0 +1,7 @@
+async function addAsync(x, y) {
+    return x + y;
+}
+
+module.exports = {
+    addAsync
+};